\name{ggiconZY-package}
\alias{ggiconZY}
\alias{ggiconZY-package}
\docType{package}
\title{Scientific Icons for ggplot2}
\description{
Reusable coordinate data and plotting helpers for embedding scientific icons
and bacterial culture plates in ggplot2 figures.
}
\seealso{
\code{\link{ggicon_names}}, \code{\link{ggicon_plot}},
\code{\link{annotation_ggicon}}, \code{\link{culture_plate_plot}}, and
\code{\link{well_plate_plot}}.
}
